gl-renderer: use fb_size and area for border
This step prepares to share the coordinate computations between
draw_output_borders() and output_get_border_damage(). The use of
weston_output is replaced with gl_output_state, so that when sharing the
code in a new function, it does not need a weston_output.

This stops the function from accessing output->current_mode and use the
gl-renderer tracked frambuffer size and compositing area instead. Not
using current_mode is a small step towards allowing gl-renderer to
render for other targets than an output.

No behavioral changes, all the values are still the same.

See the diagram in gl-renderer.h for the border areas.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>