simple-egl: Update buffer_size dimensions when starting as maximized
With commit 62ab6891db9763dd44228, 'clients/simple-egl: Handle buffer
scale and transform' we changed the way we resized the client, by
encapsulating the resize in update_buffer_geometry() function.

we didn't correct that when creating the EGL window, which might be
problematic if you attempt to start the window with different a
different state, like maximized.

Fixes 62ab6891db9763dd4

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>