input: fix surface jump test
There are two problems here, one is that the surface jump logic only makes
sense if the view remains the same.

The more important fix is that pointer coordinates are in global coordinates
and we want view coordinates, so this test was always wrong and led to an
xdg ping storm due to spurious focus changes.

Fixes 4d141a788

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>