backend-headless: clean up gl border data pointer
GL-renderer keeps the pointer for the border image data for later use.
When we destroy our cairo_surface, that data gets freed. Therefore,
reset the pointer stored by GL-renderer too, to ensure use-after-free
cannot happen.

This is not really necessary in this particular case, because the
renderer output is destroyed immediately after, so there is no chance of
UAF.

However, this is needed for code sharing purposes. Wayland-backend will
do exactly this when an output is resized. To share this code with
wayland-backend, we also need to reset the pointers. It's harmless and
more correct in the output disable path.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>