libweston/input: update view transforms when handling confined pointer motion
When the pointer is confined and we are handling motion, we need
up-to-date view transforms in order to respect the confinement. So
update view transforms in such case.

This fixes an issue in which we'd try to transform views with dirty
transforms in weston_pointer_clamp_event_to_region(), hitting an assert.

This is a leftover of d611ab24 "libweston: Update view transforms more
often".

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>