README: Reorder sections, rewrite intro + libweston
Reorder the sections into what people care about: what is it, how do I
build it, how do I run it, where are the docs for when I want to do
more?

The libweston section has been substantially rewritten to trim the
details down to what is accurate and relevant, and the introduction has
accordingly been updated to reflect the reality of Weston as a project
today.

Signed-off-by: Daniel Stone <daniels@collabora.com>